Update the library to work with `double` vectors and values instead of `float`s.

+++ src/BayesFilters/include/BayesFilters/AdditiveStateModel.h
@@ -15,11 +15,11 @@ class bfl::AdditiveStateModel : public bfl::StateModel
 public:
     virtual ~AdditiveStateModel() noexcept { };
 
-    virtual void motion(const Eigen::Ref<const Eigen::MatrixXf>& cur_states, Eigen::Ref<Eigen::MatrixXf> mot_states) override;
+    virtual void motion(const Eigen::Ref<const Eigen::MatrixXd>& cur_states, Eigen::Ref<Eigen::MatrixXd> mot_states) override;
 
-    virtual Eigen::MatrixXf getNoiseCovarianceMatrix() = 0;
+    virtual Eigen::MatrixXd getNoiseCovarianceMatrix() = 0;
 
-    virtual Eigen::MatrixXf getNoiseSample(const std::size_t num) = 0;
+    virtual Eigen::MatrixXd getNoiseSample(const std::size_t num) = 0;
 };
 
 #endif /* ADDITIVESTATEMODEL_H */
