+++ src/BayesFilters/include/BayesFilters/LinearStateModel.h
@@ -15,9 +15,9 @@ class bfl::LinearStateModel : public bfl::AdditiveStateModel
 public:
     virtual ~LinearStateModel() noexcept { };
 
-    virtual void propagate(const Eigen::Ref<const Eigen::MatrixXf>& cur_states, Eigen::Ref<Eigen::MatrixXf> mot_states) override;
+    virtual void propagate(const Eigen::Ref<const Eigen::MatrixXd>& cur_states, Eigen::Ref<Eigen::MatrixXd> mot_states) override;
 
-    virtual Eigen::MatrixXf getStateTransitionMatrix() = 0;
+    virtual Eigen::MatrixXd getStateTransitionMatrix() = 0;
 };
 
 #endif /* LINEARSTATEMODEL_H */
