+++ src/BayesFilters/src/LinearStateModel.cpp
@@ -6,7 +6,7 @@ using namespace bfl;
 using namespace Eigen;
 
 
-void LinearStateModel::propagate(const Eigen::Ref<const Eigen::MatrixXf>& cur_states, Eigen::Ref<Eigen::MatrixXf> pred_states)
+void LinearStateModel::propagate(const Eigen::Ref<const Eigen::MatrixXd>& cur_states, Eigen::Ref<Eigen::MatrixXd> pred_states)
 {
     pred_states = getStateTransitionMatrix() * cur_states;
 }
